Проработка ветки категорий

+++ backend/core/src/main/java/ru/kuzmin/secret/core/bean/CategoryServiceImpl.java
@@ -7,6 +7,7 @@ import ru.kuzmin.secret.core.dao.CategoryDao;
 import ru.kuzmin.secret.core.entity.Category;
 
 import java.util.List;
+import java.util.Objects;
 
 @Service
 public class CategoryServiceImpl implements CategoryService {
@@ -29,15 +30,69 @@ public class CategoryServiceImpl implements CategoryService {
     }
 
     @Override
+    public void delete(Long categoryId) {
+        Objects.requireNonNull(categoryId);
+        Category category = categoryDao.load(categoryId)
+                .orElseThrow(() -> new IllegalArgumentException("Category with id = " + categoryId + " not found"));
+        delete(category);
+    }
+
+    @Override
+    public void addSubcategory(Long categoryId, Long subcategoryId) {
+        Objects.requireNonNull(categoryId);
+        Objects.requireNonNull(subcategoryId);
+        Category category = categoryDao.load(categoryId)
+                .orElseThrow(() -> new IllegalArgumentException("Category with id = " + categoryId + " not found"));
+        Category subcategory = categoryDao.load(subcategoryId)
+                .orElseThrow(() -> new IllegalArgumentException("Category with id = " + subcategoryId + " not found"));
+        addSubcategory(category, subcategory);
+    }
+
+    @Override
+    public void removeSubcategory(Long categoryId, Long subcategoryId) {
+        Objects.requireNonNull(categoryId);
+        Objects.requireNonNull(subcategoryId);
+        Category category = categoryDao.load(categoryId)
+                .orElseThrow(() -> new IllegalArgumentException("Category with id = " + categoryId + " not found"));
+        Category subcategory = categoryDao.load(subcategoryId)
+                .orElseThrow(() -> new IllegalArgumentException("Category with id = " + subcategoryId + " not found"));
+        removeSubcategory(category, subcategory);
+    }
+
+    @Override
+    public List<Category> getSubcategories(Long categoryId) {
+        Objects.requireNonNull(categoryId);
+        Category category = categoryDao.load(categoryId)
+                .orElseThrow(() -> new IllegalArgumentException("Category with id = " + categoryId + " not found"));
+        return getSubcategories(category);
+    }
+
+    @Override
     public void delete(Category category) {
+        if (!getSubcategories(category).isEmpty()) {
+            throw new IllegalStateException("Category with id = " + category.getId() + " has subcategories");
+        }
         categoryDao.delete(category);
     }
 
     @Override
-    public List<Category> getSubcategories(Category category) {
-        return category == null
-                ? categoryDao.loadRootCategories()
-                : categoryDao.loadSubcategories(category);
+    public void addSubcategory(Category category, Category subcategory) {
+        subcategory.setParentId(category.getId());
+        categoryDao.update(subcategory);
     }
 
+    @Override
+    public void removeSubcategory(Category category, Category subcategory) {
+        if (Objects.equals(subcategory.getParentId(), category.getId())) {
+            subcategory.setParentId(null);
+            categoryDao.update(subcategory);
+        } else {
+            throw new IllegalStateException("Category with id = " + category + " isn't parent category");
+        }
+    }
+
+    @Override
+    public List<Category> getSubcategories(Category category)  {
+        return categoryDao.loadSubcategories(category);
+    }
 }
